backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.16.1
excel-base==1.0.4
isoweek==1.3.3
prettyconf==2.3.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
screen==1.0.1
six==1.17.0
sqlparse==0.5.3
TimeConvert==3.0.13
typing_extensions==4.15.0
tzlocal==5.3.1
xlwt==1.3.0
